renamed config object, renamed configuration directory and added stage to config file names
